+++ src/index.ts
@@ -87,6 +87,7 @@ export * from './http/ldp/PutOperationHandler';
 // HTTP/Output/Error
 export * from './http/output/error/ConvertingErrorHandler';
 export * from './http/output/error/ErrorHandler';
+export * from './http/output/error/RedirectingErrorHandler';
 export * from './http/output/error/SafeErrorHandler';
 
 // HTTP/Output/Metadata
@@ -148,10 +149,13 @@ export * from './identity/interaction/routing/InteractionRoute';
 
 // Identity/Interaction/Util
 export * from './identity/interaction/util/BaseEmailSender';
+export * from './identity/interaction/util/BaseInteractionCompleter';
 export * from './identity/interaction/util/EmailSender';
 export * from './identity/interaction/util/InteractionCompleter';
 
 // Identity/Interaction
+export * from './identity/interaction/CompletingInteractionHandler';
+export * from './identity/interaction/InteractionHandler';
 export * from './identity/interaction/SessionHttpHandler';
 
 // Identity/Ownership
@@ -166,6 +170,7 @@ export * from './identity/storage/WebIdAdapterFactory';
 
 // Identity
 export * from './identity/IdentityProviderHttpHandler';
+export * from './identity/OidcHttpHandler';
 
 // Init/Final
 export * from './init/final/Finalizable';
@@ -337,13 +342,16 @@ export * from './util/errors/BadRequestHttpError';
 export * from './util/errors/ConflictHttpError';
 export * from './util/errors/ErrorUtil';
 export * from './util/errors/ForbiddenHttpError';
+export * from './util/errors/FoundHttpError';
 export * from './util/errors/HttpError';
 export * from './util/errors/HttpErrorUtil';
 export * from './util/errors/InternalServerError';
 export * from './util/errors/MethodNotAllowedHttpError';
+export * from './util/errors/MovedPermanentlyHttpError';
 export * from './util/errors/NotFoundHttpError';
 export * from './util/errors/NotImplementedHttpError';
 export * from './util/errors/PreconditionFailedHttpError';
+export * from './util/errors/RedirectHttpError';
 export * from './util/errors/SystemError';
 export * from './util/errors/UnauthorizedHttpError';
 export * from './util/errors/UnsupportedMediaTypeHttpError';
@@ -374,6 +382,7 @@ export * from './util/locking/RedisResourceLocker';
 export * from './util/locking/ResourceLocker';
 export * from './util/locking/SingleThreadedResourceLocker';
 export * from './util/locking/WrappedExpiringReadWriteLocker';
+export * from './util/locking/VoidLocker';
 
 // Util/Templates
 export * from './util/templates/ChainedTemplateEngine';