+++ test/unit/identity/IdentityProviderHttpHandler.test.ts
@@ -9,7 +9,6 @@ import type { ProviderFactory } from '../../../src/identity/configuration/Provid
 import type { IdentityProviderHttpHandlerArgs } from '../../../src/identity/IdentityProviderHttpHandler';
 import { IdentityProviderHttpHandler } from '../../../src/identity/IdentityProviderHttpHandler';
 import type { InteractionRoute } from '../../../src/identity/interaction/routing/InteractionRoute';
-import type { InteractionCompleter } from '../../../src/identity/interaction/util/InteractionCompleter';
 import type { HttpRequest } from '../../../src/server/HttpRequest';
 import type { HttpResponse } from '../../../src/server/HttpResponse';
 import { getBestPreference } from '../../../src/storage/conversion/ConversionUtil';
@@ -17,9 +16,10 @@ import type {
   RepresentationConverter,
   RepresentationConverterArgs,
 } from '../../../src/storage/conversion/RepresentationConverter';
+import { NotFoundHttpError } from '../../../src/util/errors/NotFoundHttpError';
 import { joinUrl } from '../../../src/util/PathUtil';
 import { guardedStreamFrom, readableToString } from '../../../src/util/StreamUtil';
-import { CONTENT_TYPE, SOLID_HTTP, SOLID_META } from '../../../src/util/Vocabularies';
+import { CONTENT_TYPE, SOLID_META } from '../../../src/util/Vocabularies';
 
 describe('An IdentityProviderHttpHandler', (): void => {
   const apiVersion = '0.2';
@@ -31,7 +31,6 @@ describe('An IdentityProviderHttpHandler', (): void => {
   let providerFactory: jest.Mocked<ProviderFactory>;
   let routes: Record<'response' | 'complete' | 'error', jest.Mocked<InteractionRoute>>;
   let controls: Record<string, string>;
-  let interactionCompleter: jest.Mocked<InteractionCompleter>;
   let converter: jest.Mocked<RepresentationConverter>;
   let errorHandler: jest.Mocked<ErrorHandler>;
   let provider: jest.Mocked<Provider>;
@@ -46,7 +45,6 @@ describe('An IdentityProviderHttpHandler', (): void => {
     };
 
     provider = {
-      callback: jest.fn(),
       interactionDetails: jest.fn(),
     } as any;
 
@@ -94,8 +92,6 @@ describe('An IdentityProviderHttpHandler', (): void => {
       }),
     } as any;
 
-    interactionCompleter = { handleSafe: jest.fn().mockResolvedValue('http://test.com/idp/auth') } as any;
-
     errorHandler = { handleSafe: jest.fn(({ error }: ErrorHandlerArgs): ResponseDescription => ({
       statusCode: 400,
       data: guardedStreamFrom(`{ "name": "${error.name}", "message": "${error.message}" }`),
@@ -107,17 +103,14 @@ describe('An IdentityProviderHttpHandler', (): void => {
       providerFactory,
       interactionRoutes: Object.values(routes),
       converter,
-      interactionCompleter,
       errorHandler,
     };
     handler = new IdentityProviderHttpHandler(args);
   });
 
-  it('calls the provider if there is no matching route.', async(): Promise<void> => {
+  it('throws a 404 if there is no matching route.', async(): Promise<void> => {
     operation.target.path = joinUrl(baseUrl, 'invalid');
-    await expect(handler.handle({ request, response, operation })).resolves.toBeUndefined();
-    expect(provider.callback).toHaveBeenCalledTimes(1);
-    expect(provider.callback).toHaveBeenLastCalledWith(request, response);
+    await expect(handler.handle({ request, response, operation })).rejects.toThrow(NotFoundHttpError);
   });
 
   it('creates Representations for InteractionResponseResults.', async(): Promise<void> => {
@@ -185,39 +178,5 @@ describe('An IdentityProviderHttpHandler', (): void => {
     const result = (await handler.handle({ request, response, operation }))!;
     expect(result).toBeDefined();
     expect(JSON.parse(await readableToString(result.data!))).toEqual({ apiVersion, authenticating: true, controls });
-  });
-
-  it('errors for InteractionCompleteResults if no oidcInteraction is defined.', async(): Promise<void> => {
-    operation.target.path = joinUrl(baseUrl, '/idp/routeComplete');
-    operation.method = 'POST';
-
-    const error = expect.objectContaining({
-      statusCode: 400,
-      message: 'This action can only be performed as part of an OIDC authentication flow.',
-      errorCode: 'E0002',
-    });
-    await expect(handler.handle({ request, response, operation })).rejects.toThrow(error);
-    expect(routes.complete.handleOperation).toHaveBeenCalledTimes(1);
-    expect(routes.complete.handleOperation).toHaveBeenLastCalledWith(operation, undefined);
-    expect(interactionCompleter.handleSafe).toHaveBeenCalledTimes(0);
-  });
-
-  it('calls the interactionCompleter for InteractionCompleteResults and redirects.', async(): Promise<void> => {
-    operation.target.path = joinUrl(baseUrl, '/idp/routeComplete');
-    operation.method = 'POST';
-    operation.body = new BasicRepresentation('value', 'text/plain');
-    const oidcInteraction = { session: { accountId: 'account' }, prompt: {}} as any;
-    provider.interactionDetails.mockResolvedValueOnce(oidcInteraction);
-    const result = (await handler.handle({ request, response, operation }))!;
-    expect(result).toBeDefined();
-    expect(routes.complete.handleOperation).toHaveBeenCalledTimes(1);
-    expect(routes.complete.handleOperation).toHaveBeenLastCalledWith(operation, oidcInteraction);
-    expect(operation.body?.metadata.contentType).toBe('application/json');
-
-    expect(interactionCompleter.handleSafe).toHaveBeenCalledTimes(1);
-    expect(interactionCompleter.handleSafe).toHaveBeenLastCalledWith({ request, webId: 'webId' });
-    const location = await interactionCompleter.handleSafe.mock.results[0].value;
-    expect(result.statusCode).toBe(302);
-    expect(result.metadata?.get(SOLID_HTTP.terms.location)?.value).toBe(location);
   });
 });